Implement UserModel class, simplify UserProfileStore
